requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
billiard==3.6.4.0
build==0.8.0
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.1.1
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.13
django-annoying==0.10.6
django-celery-beat==2.2.1
django-ckeditor==6.3.2
django-cors-headers==3.11.0
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-taggit==2.1.0
django-templated-mail==1.1.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
drf-yasg==1.20.0
flake8==5.0.4
flower==1.0.0
humanize==4.0.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
pep517==0.12.0
phonenumberslite==8.12.46
Pillow==9.1.0
pip-compile-multi==2.4.5
pip-tools==6.6.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pydantic==1.9.2
pyflakes==2.5.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
toposort==1.7
tornado==6.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
